# lib/api-zod/src/generated/types/authResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Student Mentor AI API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AuthResultRole } from './authResultRole';

export interface AuthResult {
  success: boolean;
  role: AuthResultRole;
  token: string;
  message?: string;
}