# lib/api-zod/src/generated/types/factorAnalysis.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Student Mentor AI API specification
 * OpenAPI spec version: 0.1.0
 */
import type { FactorGroup } from './factorGroup';

export interface FactorAnalysis {
  factor: string;
  groups: FactorGroup[];
}