# lib/api-zod/src/generated/types/predictionInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Student Mentor AI API specification
 * OpenAPI spec version: 0.1.0
 */

export interface PredictionInput {
  hoursStudied: number;
  attendance: number;
  sleepHours: number;
  previousScores: number;
  tutoringSessions: number;
  physicalActivity: number;
  parentalInvolvement: string;
  accessToResources: string;
  motivationLevel: string;
  internetAccess: string;
  familyIncome: string;
  teacherQuality: string;
  peerInfluence: string;
  extracurricularActivities: string;
}